Add a partial-skolemization (bound-variable renaming) module in `goto-programs`.
That allows us to do a bound-variables renaming early in `process-goto-programs`,
which then allows us to not have problems with clashing identifiers later on.

This performs only partial-skolemisation (doesn't eliminate the quantifier, but
renames the bound variables) because for the correct operation of VCCs we need
to preserve the exists statement.