Add array type to sort conversion
This is sufficient for supporting the conversion of the generated
expressions for non-deterministically initialised arrays. However it is
not yet straight forward to regression test without support for the
conversion of further kinds of expression. This is because the
initialisation would be sliced away if the result of the initialisation
is not read back later on.