Use `get_bool_option` to get `unwinding-assertions` option
The `optionst` class essentially uses tri-state logic for booleans,
where an option may be true, false or unset. `is_set` will return `true`
for both the true and false options. Therefore we need `get_bool_option`
to get the state of whether the option is set to true or false.